Add MSVC compatibility and improve Windows support
This commit introduces changes for better compatibility with Microsoft Visual C++ (MSVC) and overall improvements for Windows platform. The CMake build configuration was updated to include specific optimization flags and warnings for MSVC on top of existing flags for GCC and Clang. The `logger.h` now correctly includes `<typeinfo>` and handles class name demangling when used with MSVC, rather than just GCC/Clang.

In `color.h`, color formatting was disabled for Windows, as `fmt::text_style` isn't yet supported on this platform. A line ending configuration for Clang was also added to ensure consistent LF line endings across platforms. The changes are made to extend the compatibility to a broader range of systems and for better portability.