Implement rolling log file functionality
Added a new feature into the FileLogger that rolls over the file when a specified size is reached. A backup file is created when rolling over, using the current timestamp as part of the backup file name. Also, a new example of this feature's usage was added along with its own CMakeList for building. Finally, a minor modification was made to the include path in LogManager.h.