Update formatting calls to use the new LogFormat methods
The log formatting functions in different logger classes have been updated to use a new Get method from the LogFormat class instead of the previous DEFAULT constant. This adjustment will allow for more flexibility in log formatting as the format can now be changed dynamically using the Set method in the LogFormat class.